test(a11y): break out accessibility tests into separate files for each chart type
